Add a PickingResult object in plot and inherit from it in plot3d
